+++ database/migrations/2026_02_10_300_create_pembayarans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('pembayarans', function (Blueprint $table) {
             $table->id();
             $table->foreignId('transaksi_id')->constrained();
-            $table->decimal('jumlah_bayar', 10, 2);
+            $table->integer('jumlah_bayar');
             $table->timestamps();
         });
     }
